\name{DBI.RPgSQL-internal}
\alias{.conflicts.OK}
\alias{.__C__PgSQLDriver}
\alias{.__C__Id}
\alias{.__C__PgSQLConnection}
\alias{.__C__PgSQLObject}
\alias{.__C__PgSQLResult}
\alias{.__M__[}
\alias{.__M__[<-}
\alias{.__M__dbClearResult}
\alias{.__M__dbCommit}
\alias{.__M__dbCallProc}
\alias{.__M__dbConnect}
\alias{.__M__dbDataType}
\alias{.__M__dbDisconnect}
\alias{.__M__dbExistsTable}
\alias{.__M__dbGetException}
\alias{.__M__dbGetInfo}
\alias{.__M__dbColumnInfo}
\alias{.__M__dbGetQuery}
\alias{.__M__dbGetStatement}
\alias{.__M__dbHasCompleted}
\alias{.__M__dbListConnections}
\alias{.__M__dbListTables}
\alias{.__M__dbListFields}
\alias{.__M__dbListResults}
\alias{.__M__dbReadTable}
\alias{.__M__dbRemoveTable}
\alias{.__M__dbRollback}
\alias{.__M__dbSendQuery}
\alias{.__M__dbUnloadDriver}
\alias{.__M__dbWriteTable}
\alias{.__M__fetch}
\alias{.__M__format}
\alias{.__M__isSQLKeyword}
\alias{.__M__make.db.names}
\alias{.__M__print}
\alias{.__M__show}
\title{Version 4 meta-data in the DBI.RPgSQL package}
\description{
  Internal object to Version 4 classes and methods.
}

\keyword{internal}
\keyword{interface}
\keyword{database}

% vim: syntax=tex
